# Improve Performance for manage.py Commands or Alternative (non-daphne) ASGI Servers