--- a/channels/apps.py
+++ b/channels/apps.py
@@ -1,10 +1,18 @@
-# We import this here to ensure the reactor is installed very early on
-# in case other packages accidentally import twisted.internet.reactor
-# (e.g. raven does this).
-import daphne.server
-from django.apps import AppConfig
+import sys
 
-assert daphne.server  # pyflakes doesn't support ignores
+if (
+    len(sys.argv) >= 2
+    and sys.argv[0].endswith("manage.py")
+    and sys.argv[1] == "runserver"
+) or sys.argv[0].endswith("daphne"):
+    # We import this here to ensure the reactor is installed very early on
+    # in case other packages accidentally import twisted.internet.reactor
+    # (e.g. raven does this).
+    import daphne.server
+
+    assert daphne.server  # pyflakes doesn't support ignores
+
+from django.apps import AppConfig
 
 
 class ChannelsConfig(AppConfig):