This improves speed and memory usage of `manage.py` commands other than `runserver`. It also improves performance when using an alternative ASGI server like "uvicorn" in production. In these cases `daphne` is not needed at all, and we can avoid the import cost of `daphne`.

## Reproduction

Create a new Django project, install channels and memray, apply migrations.

    django-admin startproject mysite
    cd mysite
    pip install Django channels memray
    ./manage.py migrate

Run `manage.py showmigrations` using memray.

    python -m memray run manage.py showmigrations
    python -m memray flamegraph memray-manage.py.253077.bin

Open HTML report and view total memory usage and memory allocations. Run `time ./manage.py showmigrations` 10 times and calculate average execution time.

Apply the changes from this PR, and calculate the metrics again.

## Results:

| Metric                | Before   | After    | Change       |
| --------------------- | -------- | -------- | ------------ |
| Memray - Total Memory | 32.5 MB | 17.9 MB | -45% |
| Memray - Allocations  | 10991    | 3992    | -64%      |
| Execution Time         | 0.322s  | 0.189s  | -41%     |

As you see, this PR can reduce memory usage up to **45%** and execution time up to **41%**.

## Package Versions

```
Django==4.0.4
channels==3.0.4
daphne==3.0.2
asgiref==3.5.0
memray==1.0.3
```