@carltongibson 2022-08-08
Closing in favour of #1897, which will be in Channels v4 around early-September.